requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
appdirs==1.4.4
asgiref==3.8.1
attrs==24.2.0
beautifulsoup4==4.12.3
blessed==1.20.0
blinker==1.8.2
bs4==0.0.2
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
cssselect==1.2.0
Deprecated==1.2.14
distro==1.9.0
fake-useragent==1.5.1
Flask==3.0.3
Flask-Cors==5.0.0
Flask-Limiter==3.8.0
frozendict==2.4.4
frozenlist==1.4.1
gunicorn==23.0.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
html5lib==1.1
httpcore==1.0.5
httpx==0.27.2
Hypercorn==0.17.3
hyperframe==6.0.1
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.4.5
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.5.0
Js2Py==0.74
json5==0.9.25
limits==3.13.0
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
multidict==6.1.0
multitasking==0.0.11
nest-asyncio==1.6.0
numpy==2.1.1
openai==1.45.1
ordered-set==4.1.0
outcome==1.3.0.post0
packaging==24.1
pandas==2.2.2
parse==1.20.2
peewee==3.17.6
pillow==10.4.0
platformdirs==4.3.3
priority==2.0.0
psutil==6.0.0
pydantic==2.9.1
pydantic_core==2.23.3
pyee==11.1.1
Pygments==2.18.0
pyjsparser==2.7.1
pyppeteer==2.0.0
pyquery==2.0.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
regex==2024.9.11
requests==2.32.3
requests-html==0.10.0
rich==13.8.1
selenium==4.24.0
selenium-stealth==1.0.6
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
tabulate==0.9.0
tavily-python==0.4.0
tiktoken==0.7.0
tqdm==4.66.5
trio==0.26.2
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
undetected-chromedriver==3.5.5
urllib3==1.26.20
uvicorn==0.30.6
w3lib==2.2.1
wcwidth==0.2.13
webencodings==0.5.1
websocket-client==1.8.0
websockets==10.4
Werkzeug==3.0.4
wrapt==1.16.0
wsproto==1.2.0
yarl==1.11.1
yfinance==0.2.43
zipp==3.20.2
